development: add command to configure SiK radio
This is a command that allows to set AT parameters of SiK radios
connected to a drone.

For example, this can be used to set the ID of a radio allowing it to be
paired to another ground radio. This can be used for configuration at
initial setup but also during a mission to specifically swap to a new
ground radio.

Signed-off-by: Julian Oes <julian@oes.ch>